Revert incorrect previous package rename
