+++ cf-ops-automation-sample-broker/src/test/java/com/orange/oss/cloudfoundry/broker/opsautomation/ondemandbroker/sample/ServiceBindingTest.java
@@ -1,4 +1,4 @@
-package com.orange.oss.cloudfoundry.broker.opsautomation.ondemandbroker.cloudflare;
+package com.orange.oss.cloudfoundry.broker.opsautomation.ondemandbroker.sample;
 
 
 import io.restassured.RestAssured;
